requirements.txt
# backend.
openai>=1.0.0
python-dotenv>=1.0.0
flask>=2.3.0
requests>=2.32.3
mysql-connector-python>=9.0.0
langchain-openai>=0.3.0